vpn/netstate: run the linux route monitor for the whole process lifetime
The netlink route-change monitor moves from routeState (started by every
EnableClientRoutes, stopped by teardown) into Manager: Start(ctx) launches
it once and it lives until ctx is cancelled. routeState becomes a pure
value-holder — its mutex and monitor lifecycle fields are gone, everything
is guarded by Manager.mu, so reconcile and Enable/Disable transitions
serialise on the same lock. The subscription loop lives in
manager_linux.go, the route-diff logic stays in routes_linux.go.

An initial subscribe failure no longer disables staleness tracking until
the next enable: the monitor starts with a dead subscription and keeps
re-subscribing with backoff.
